HCK-7465: resolve comments
